(완료) 출력 시 1, 11 -> ACE 로, 10-> J or Q or K 로 변환.

+++ src/main/java/domain/PlayGame.java
@@ -157,9 +157,9 @@ public class PlayGame {
     }
 
     private void printFinalResult() {
-        System.out.println('\n' + dealer.print() + " - 결과 :" + dealer.cardSum());
+        System.out.println('\n' + dealer.print() + " - 결과 :" + (int)dealer.cardSum());
         for (Player player : players) {
-            System.out.println(player.print() + " - 결과 :" + player.cardSum());
+            System.out.println(player.print() + " - 결과 :" + (int)player.cardSum());
         }
         printFinalScore();
     }
